documenting TrajectoryCalculator
