+++ src/main/java/com/team1816/lib/auto/paths/AutoPath.java
@@ -6,7 +6,6 @@ import com.team1816.season.configuration.Constants;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.trajectory.Trajectory;
-import com.team1816.lib.motion.Trajectories;
 
 import java.util.ArrayList;
 import java.util.List;
