--- src/main/resources/trajectories/README.md
@@ -1,5 +1,0 @@
-This is a directory anchor for the src/main/resources/trajectories directory.
-
-Said directory holds the JSON stringified states of auto trajectories that are loaded on startup.
-
-Populate this directory with ``gradlew calcTrajectories``.
